set service name and distro metadata via SPI
Fixes #40

Previously, we were clobbering any resource attributes set via environment
variable (OTEL_RESOURCE_ATTRIBUTES) or properties (otel.resource.attributes).
This change uses the OTel Java autoconfiguration SPI to set the service
name and distro metadata fields via a ResourceProvider. This is akin to the
pattern for configuring the trace provider via an SdkTracerProviderConfigurer
that we're already using to customize tracing.

We're keeping the clobber of otel.exporter.otlp.(headers|endpoint) properties
in HoneycombAgent.configure() because those properties shouldn't be set to
anything else when sending to Honeycomb.

Co-authored-by: Vera Reynolds <verareynolds@honeycomb.io>