share a collector image in the smoke tests
The Collector is configured the same way in each smoke test environment,
so make a single directory for its build context and reference that
directory in the smoke tests' docker-compose configs. NOTE: the running
collector container for each environment will remain *NOT* shared. This
is to maintain collector setup only once.