Smoke tests for SDK-only traces (& duplication reduction) (#197)
* tell gradle to place the JARs directly in build-artifacts

... instead of the JARs get sprinkled throughout the subprojects in
build/libs.

Wires up the smoke tests to know that they depend on JARs from the
build-artifacts directory and how to copy them to the locations and
filenames the Dockerfiles expect.

* persist build-artifacts from the working dir instead of HOME

* share a collector image in the smoke tests

The Collector is configured the same way in each smoke test environment,
so make a single directory for its build context and reference that
directory in the smoke tests' docker-compose configs. NOTE: the running
collector container for each environment will remain *NOT* shared. This
is to maintain collector setup only once.

* strict string comparison instead of grepping

This means the collector must have been restarted before rerunning a
smoke test to flush the JSON serialization of received trace data.

* chainsaw surgery for de-dup

== APPS ==

One app.dockerfile with multi-stage build targets. docker-compose.yml
updated to target those ..uh.. targets.

== BATS ==

One bats.dockerfile with all the bats files in it. docker-compose.yml
updated with individual bats services and a specific CMD given for which
bats file to run.

== MAKEFILE ==

Updated to use the one docker-compose.yml and the --project-name option
for isolating the different smokey environments. JAR files used tucked
into ./smoke-tests/jars/ for good hygiene.

* move common test helpers to test_helpers

We can do this now that the container built for BATS has a single build
context (in other words, all the bats are in one directory.

* smoke testing on a VM

Move to machine executor for smoke tests. It's slower, but the
docker services running there have more options, like volume
mounts.

* emojis make everything better

Co-authored-by: Robb Kidd <robbkidd@honeycomb.io>