maint: add smoke tests for grpc/http exporter protocols for agent only (#391)
## Which problem is this PR solving?

- closes #287 

## Short description of the changes

- add smoke tests for the app using only the java agent, similar to the
tests for the SDK, for grpc and http/protobuf protocols
- there is a TODO added into the http test, as metrics are not currently
available with http/protobuf protocol (see #293)

I initially [started this for the
app-agent-manual](https://github.com/honeycombio/honeycomb-opentelemetry-java/compare/jamie.smoke-agent-manual-protocols?expand=1),
which is a mix of agent and manual instrumentation, but thought it was
sufficient to only include the agent-only example. The test is for
exporting with the specific protocol, which should have no impact on
whether it's auto or manual as long as metrics are exported for the
agent (which they are).